Fix Refactor errors
