+++ src/main/java/seedu/address/logic/commands/DeleteCommand.java
@@ -2,8 +2,8 @@ package seedu.address.logic.commands;
 
 import seedu.address.commons.core.Messages;
 import seedu.address.commons.core.UnmodifiableObservableList;
-import seedu.address.model.task.ReadOnlyPerson;
-import seedu.address.model.task.UniquePersonList.PersonNotFoundException;
+import seedu.address.model.person.UniquePersonList.PersonNotFoundException;
+import seedu.address.model.person.ReadOnlyPerson;
 
 /**
  * Deletes a person identified using it's last displayed index from the address book.
