+++ src/main/java/seedu/address/model/person/ReadOnlyTask.java
@@ -1,4 +1,4 @@
-package seedu.address.model.task;
+package seedu.address.model.person;
 
 import java.util.Date;
