+++ src/test/java/seedu/address/testutil/AddressBookBuilder.java
@@ -5,6 +5,7 @@ import seedu.address.model.tag.Tag;
 import seedu.address.model.AddressBook;
 import seedu.address.model.person.Person;
 import seedu.address.model.person.UniquePersonList;
+import seedu.address.model.person.UniquePersonList.DuplicatePersonException;
 
 /**
  * A utility class to help with building Addressbook objects.
@@ -19,7 +20,7 @@ public class AddressBookBuilder {
         this.addressBook = addressBook;
     }
 
-    public AddressBookBuilder withPerson(Person person) throws UniquePersonList.DuplicatePersonException {
+    public AddressBookBuilder withPerson(Person person) throws DuplicatePersonException {
         addressBook.addPerson(person);
         return this;
     }
